Data Analysis Practice
To study how income varies across different categories of college majors. You will be using data from a study of recent college graduates. Make sure to use good practices that you have learned so far in this course and previous courses in the specialization.

Results: 
We analyzed the data records for 15 different college majors and how it associate with the income of people. It was found that choosing a college education has a significant impact on the income of a person. But, selecting the right major can lead to higher income. The study revealed that more 25% of majors can leave earning that are below the median income. If you had to choose the major that gives you the highest income, then you would choose Business. This is no surprise that business leaders are paid the most since they run the companies. On the contrary, going for a major like Interdisciplinary will give you the lowest income.